+++ src/models/Point.java
@@ -1,16 +1,16 @@
 package models;
 
 public class Point {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -18,7 +18,7 @@ public class Point {
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
